Fires callback if event is triggered within the macro's `element`
Before this change, events where not triggered when any element in the DOM was
focused (other than the macro's provided `element` scope). This led to
unexpected/undesired behavior of needing to click off of a link after it was
clicked, out of an input when it was focused (even with `isDisabledOnInput` set
to `false`).